rearrange tests and add more spread tests
